refactor: split region map settlements and province label parts (#3878)
Decompose settlement marker painting into town/port vs warp-zone parts and
split province political label compute from paint for Phase 3 modularization.
