+++ app/lib/features/game/flame/region_map/region_map_component.dart
@@ -35,11 +35,13 @@ part 'region_map_component_render_core_base_tiles_sea.dart';
 part 'region_map_component_render_core_base_tiles_land.dart';
 part 'region_map_component_render_core_transport_feature.dart';
 part 'region_map_component_render_core_overlays.dart';
-part 'region_map_component_render_political_labels_province.dart';
+part 'region_map_component_render_political_labels_province_compute.dart';
+part 'region_map_component_render_political_labels_province_paint.dart';
 part 'region_map_component_render_political_labels_sea.dart';
 part 'region_map_component_render_political_borders.dart';
 part 'region_map_component_render_markers_selection.dart';
-part 'region_map_component_render_markers_settlements.dart';
+part 'region_map_component_render_markers_settlements_towns.dart';
+part 'region_map_component_render_markers_settlements_warp.dart';
 part 'region_map_component_render_markers_units.dart';
 
 final _log = packageLogger();
